Fix Kotlin extensions
